Guessing Game
A game where the computer picks a random number from 1 to 1,023, and the user must guess the number. The user is told if there guess is too high or too low, which can help the player win faster. The computer keeps track of the number of distinct guesses the user makes, and tells the user after every round and at the end of the game. At the end of the game, the computer also finds the minimum number of guesses need to find the number using a specific search algorithm. It compares this with the user's performance, and encourages the player to discover what the method is. There are also various quality of life elements, such as telling the player when there input is invalid or not logical, a short surprise message if the user manages to win with one guess, and several lines that correct for grammar. Enjoy!!